requirements.txt
flask
flask-cors
requests
pdfminer.six
gunicorn